+++ server/lib/postgres.js
@@ -1,0 +1,13 @@
+const keys = require('../keys');
+
+// Postgres Client Setup
+const { Pool } = require('pg');
+const pgClient = new Pool({
+    user: keys.pgUser,
+    host: keys.pgHost,
+    database: keys.pgDatabase,
+    password: keys.pgPassword,
+    port: keys.pgPort
+});
+
+module.exports = pgClient;
