+++ server/lib/serverHelperFunctions.js
